###Description: Implement description/ELI5 regarding the sentiment towards the project ###How I implemented: A file that contains all the chatGPT methods and a file that scrapes hackernews. ###Other I am currently working on the categorization wihtin the same branch. Sorry. But I am still struggleing with the types
In ScrapeRepoCategroies there is still a problem with the types (not eslint confirm) but since we need the functionality I will commit it know and modify it later. But I am just not able to to write the correct types for the complex response. There are prototypes within the types folder.